+++ src/main.tsx
@@ -6,11 +6,11 @@ import '@/index.css';
 import { initConfig } from '@/config';
 import { initTheme } from '@common/utils/theme';
 
+initTheme();
+
 registerSW();
 
 initConfig().then(async () => {
-  initTheme();
-
   const { default: App } = await import('./App');
 
   createRoot(document.getElementById('root')!).render(